Move Status and StatusOr to google/cloud.
We will be using these classes in all APIs, not just Google Cloud
Storage.
