+++ google/cloud/status.h
@@ -12,17 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_STATUS_H_
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_STATUS_H_
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STATUS_H_
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STATUS_H_
 
-#include "google/cloud/storage/version.h"
+#include "google/cloud/version.h"
 #include <iostream>
 #include <tuple>
 
 namespace google {
 namespace cloud {
-namespace storage {
-inline namespace STORAGE_CLIENT_NS {
+inline namespace GOOGLE_CLOUD_CPP_NS {
 /// Well-known status codes with `grpc::StatusCode`-compatible values.
 ///    https://grpc.io/grpc/cpp/classgrpc_1_1_status.html
 enum StatusCode {
@@ -110,28 +109,8 @@ class RuntimeStatusError : public std::runtime_error {
   Status status_;
 };
 
-/**
- * Report checksum mismatches as exceptions.
- */
-class HashMismatchError : public std::ios_base::failure {
- public:
-  explicit HashMismatchError(std::string const& msg, std::string received,
-                             std::string computed)
-      : std::ios_base::failure(msg),
-        received_hash_(std::move(received)),
-        computed_hash_(std::move(computed)) {}
-
-  std::string const& received_hash() const { return received_hash_; }
-  std::string const& computed_hash() const { return computed_hash_; }
-
- private:
-  std::string received_hash_;
-  std::string computed_hash_;
-};
-
-}  // namespace STORAGE_CLIENT_NS
-}  // namespace storage
+}  // namespace GOOGLE_CLOUD_CPP_NS
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_STATUS_H_
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STATUS_H_
