+++ google/cloud/storage/CMakeLists.txt
@@ -159,8 +159,6 @@ add_library(storage_client
             internal/service_account_requests.cc
             internal/signed_url_requests.h
             internal/signed_url_requests.cc
-            internal/throw_status_delegate.h
-            internal/throw_status_delegate.cc
             lifecycle_rule.h
             lifecycle_rule.cc
             list_buckets_reader.h
@@ -199,9 +197,6 @@ add_library(storage_client
             service_account.cc
             signed_url_options.h
             signed_url_options.cc
-            status.h
-            status.cc
-            status_or.h
             storage_class.h
             upload_options.h
             version.h
@@ -331,7 +326,6 @@ if (BUILD_TESTING)
         internal/retry_resumable_upload_session_test.cc
         internal/service_account_requests_test.cc
         internal/signed_url_requests_test.cc
-        internal/throw_status_delegate_test.cc
         lifecycle_rule_test.cc
         list_buckets_reader_test.cc
         list_objects_reader_test.cc
@@ -348,7 +342,6 @@ if (BUILD_TESTING)
         retry_policy_test.cc
         service_account_test.cc
         signed_url_options_test.cc
-        status_or_test.cc
         storage_class_test.cc
         storage_client_options_test.cc
         storage_version_test.cc
