+++ google/cloud/storage/internal/access_control_common.h
@@ -16,9 +16,9 @@
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_ACCESS_CONTROL_COMMON_H_
 
 #include "google/cloud/optional.h"
+#include "google/cloud/status.h"
 #include "google/cloud/storage/internal/common_metadata.h"
 #include "google/cloud/storage/internal/nljson.h"
-#include "google/cloud/storage/status.h"
 #include <utility>
 
 namespace google {
