+++ google/cloud/storage/internal/raw_client.h
@@ -15,6 +15,8 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_RAW_CLIENT_H_
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_RAW_CLIENT_H_
 
+#include "google/cloud/status.h"
+#include "google/cloud/status_or.h"
 #include "google/cloud/storage/bucket_metadata.h"
 #include "google/cloud/storage/client_options.h"
 #include "google/cloud/storage/internal/bucket_acl_requests.h"
@@ -30,8 +32,6 @@
 #include "google/cloud/storage/oauth2/credentials.h"
 #include "google/cloud/storage/object_metadata.h"
 #include "google/cloud/storage/service_account.h"
-#include "google/cloud/storage/status.h"
-#include "google/cloud/storage/status_or.h"
 
 namespace google {
 namespace cloud {
