+++ google/cloud/storage/list_objects_reader.h
@@ -15,6 +15,7 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_LIST_OBJECTS_READER_H_
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_LIST_OBJECTS_READER_H_
 
+#include "google/cloud/status_or.h"
 #include "google/cloud/storage/internal/object_requests.h"
 #include "google/cloud/storage/internal/raw_client.h"
 #include "google/cloud/storage/status_or.h"
