+++ google/cloud/storage/object_access_control.h
@@ -15,11 +15,11 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_OBJECT_ACCESS_CONTROL_H_
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_OBJECT_ACCESS_CONTROL_H_
 
+#include "google/cloud/status_or.h"
 #include "google/cloud/storage/internal/access_control_common.h"
 #include "google/cloud/storage/internal/common_metadata.h"
 #include "google/cloud/storage/internal/nljson.h"
 #include "google/cloud/storage/internal/patch_builder.h"
-#include "google/cloud/storage/status_or.h"
 
 namespace google {
 namespace cloud {
