+++ google/cloud/storage/object_rewriter.cc
@@ -14,7 +14,7 @@
 
 #include "google/cloud/storage/object_rewriter.h"
 #include "google/cloud/storage/internal/raw_client.h"
-#include "google/cloud/storage/internal/throw_status_delegate.h"
+#include "google/cloud/internal/throw_delegate.h"
 
 namespace google {
 namespace cloud {
