+++ google/cloud/storage/retry_policy.h
@@ -17,7 +17,7 @@
 
 #include "google/cloud/internal/backoff_policy.h"
 #include "google/cloud/internal/retry_policy.h"
-#include "google/cloud/storage/status.h"
+#include "google/cloud/status.h"
 
 namespace google {
 namespace cloud {
