--- google/cloud/storage/status.cc
@@ -1,36 +1,0 @@
-// Copyright 2018 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "google/cloud/storage/status.h"
-#include <sstream>
-
-namespace google {
-namespace cloud {
-namespace storage {
-inline namespace STORAGE_CLIENT_NS {
-namespace {
-std::string StatusWhat(Status const& status) {
-  std::ostringstream os;
-  os << status;
-  return std::move(os).str();
-}
-}  // namespace
-
-RuntimeStatusError::RuntimeStatusError(Status status)
-    : std::runtime_error(StatusWhat(status)), status_(std::move(status)) {}
-
-}  // namespace STORAGE_CLIENT_NS
-}  // namespace storage
-}  // namespace cloud
-}  // namespace google
