+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -59,7 +59,6 @@ storage_client_unit_tests = [
     "internal/retry_resumable_upload_session_test.cc",
     "internal/service_account_requests_test.cc",
     "internal/signed_url_requests_test.cc",
-    "internal/throw_status_delegate_test.cc",
     "lifecycle_rule_test.cc",
     "list_buckets_reader_test.cc",
     "list_objects_reader_test.cc",
@@ -76,7 +75,6 @@ storage_client_unit_tests = [
     "retry_policy_test.cc",
     "service_account_test.cc",
     "signed_url_options_test.cc",
-    "status_or_test.cc",
     "storage_class_test.cc",
     "storage_client_options_test.cc",
     "storage_version_test.cc",
